@@ -15,9 +15,12 @@
     <description>A spring boot application starter with Mysql, vuejs, webpack and bulma and beufy</description>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <timestamp>${maven.build.timestamp}</timestamp>
         <humanize.version>1.2.2</humanize.version>
+        <!--        <maven.compiler.source>17</maven.compiler.source>-->
+        <!--        <maven.compiler.target>17</maven.compiler.target>-->
+        <!--        <maven.compiler.release>17</maven.compiler.release>-->
         <maven.build.timestamp.format>yyyy-MM-dd.HH.mm.ss</maven.build.timestamp.format>
     </properties>
     <repositories>
@@ -333,10 +391,9 @@
             </plugin>
             <plugin>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.1</version>
                 <configuration>
-                    <source>${maven.compiler.source}</source>
-                    <target>${maven.compiler.target}</target>
+                    <!--                    <source>1.9</source>-->
+                    <!--                    <target>1.9</target>-->
                     <compilerArguments>
                         <processor>org.hibernate.jpamodelgen.JPAMetaModelEntityProcessor</processor>
                     </compilerArguments>
