@@ -1,10 +1,13 @@
 package com.ogoma.hr_core.boundaries.documents.services;
 
 import com.ogoma.hr_core.boundaries.documents.entities.Bucket;
+import io.minio.BucketExistsArgs;
+import io.minio.MakeBucketArgs;
 import io.minio.MinioClient;
+import io.minio.RemoveBucketArgs;
 import io.minio.errors.*;
 import org.springframework.stereotype.Service;
-import org.xmlpull.v1.XmlPullParserException;
+
 
 import java.io.IOException;
 import java.security.InvalidKeyException;
@@ -20,18 +23,21 @@ public class BucketsService {
         this.minioClient = minioClient;
     }
 
-    public void createBucket(String name) throws IOException, InvalidKeyException, NoSuchAlgorithmException, InsufficientDataException, InvalidResponseException, ErrorResponseException, NoResponseException, InvalidBucketNameException, XmlPullParserException, InternalException, RegionConflictException {
-        if (!minioClient.bucketExists(name)) {
-            minioClient.makeBucket(name);
+    public void createBucket(String name) throws Exception  {
+        BucketExistsArgs bucketExistsArgs= BucketExistsArgs.builder().bucket(name).build();
+        if (!minioClient.bucketExists(bucketExistsArgs)) {
+            MakeBucketArgs makeBucketArgs= MakeBucketArgs.builder().bucket(name).build();
+            minioClient.makeBucket(makeBucketArgs);
         }
     }
 
-    public void deleteBucket(String name) throws IOException, InvalidKeyException, NoSuchAlgorithmException, InsufficientDataException, InvalidResponseException, InternalException, NoResponseException, InvalidBucketNameException, XmlPullParserException, ErrorResponseException {
-        if (!minioClient.bucketExists(name)) {
-            minioClient.removeBucket(name);
+    public void deleteBucket(String name) throws Exception    {
+        if (!minioClient.bucketExists(BucketExistsArgs.builder().bucket(name).build())) {
+            RemoveBucketArgs removeBucketArgs = RemoveBucketArgs.builder().bucket(name).build();
+            minioClient.removeBucket(removeBucketArgs);
         }
     }
-    public List<Bucket> getBuckets() throws IOException, InvalidKeyException, NoSuchAlgorithmException, InsufficientDataException, InvalidResponseException, InternalException, NoResponseException, InvalidBucketNameException, XmlPullParserException, ErrorResponseException {
-        return minioClient.listBuckets().stream().map(x -> new Bucket(x.name(), x.creationDate(), null)).collect(Collectors.toList());
+    public List<Bucket> getBuckets() throws Exception   {
+        return minioClient.listBuckets().stream().map(x -> new Bucket(x.name(),null, null)).collect(Collectors.toList());
     }
 }
