@@ -1,7 +1,17 @@
 package com.ogoma.hr_core;
 
 
+import com.amazonaws.ClientConfiguration;
+import com.amazonaws.auth.AWSStaticCredentialsProvider;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.client.builder.AwsClientBuilder;
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import io.minio.MinioClient;
+import org.apache.http.conn.ssl.NoopHostnameVerifier;
+import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
+import org.apache.http.conn.ssl.TrustSelfSignedStrategy;
+import org.apache.http.ssl.SSLContexts;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -14,7 +24,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.client.RestTemplate;
 
 import java.util.concurrent.Executor;
-
 @SpringBootApplication
 @EnableJpaAuditing
 @EnableAsync
@@ -36,15 +45,19 @@ public class VueStarterApplication {
     }
 
     @Bean
-    MinioClient minioClient(@Value("${minio.end_point}") String endPoint,
-                            @Value("${minio.access_key}") String accessKey,
-                            @Value("${minio.secret_key}") String secretKey) throws Exception {
-        return new MinioClient(
-                endPoint,
-                accessKey,
-                secretKey
-        );
+    MinioClient minioClient(@Value("${minio.end.point}") String endPoint,
+                            @Value("${minio.access.key}") String accessKey,
+                            @Value("${minio.secret.key}") String secretKey) throws Exception {
+
+        var client = MinioClient.builder()
+                .credentials(accessKey, secretKey)
+                .region("eu-frankfurt-1")
+                .endpoint(endPoint).build();
+        return client;
+
+
     }
+
     @Bean(name = "threadPoolTaskExecutor")
     public Executor getAsyncExecutor() {
         ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
@@ -53,6 +66,27 @@ public class VueStarterApplication {
         return executor;
     }
 
+//    @Bean
+//    AmazonS3 s3client (
+//            @Value("${minio.access.key}") String accessKey,
+//            @Value("${minio.secret.key}") String secretKey
+//    ) throws Exception {
+//        var cred = new BasicAWSCredentials(
+//                accessKey,
+//                secretKey
+//        );
+//        ClientConfiguration clientConfiguration = new ClientConfiguration();
+//        SSLConnectionSocketFactory scsf =new  SSLConnectionSocketFactory(
+//                SSLContexts.custom().loadTrustMaterial(null, new TrustSelfSignedStrategy()).build(),
+//                NoopHostnameVerifier.INSTANCE);
+//        clientConfiguration.getApacheHttpClientConfig().setSslSocketFactory(scsf);
+//        return AmazonS3ClientBuilder
+//                .standard()
+//                .withCredentials(new AWSStaticCredentialsProvider(cred))
+//                .withRegion("us-east-2")
+//                .withClientConfiguration(clientConfiguration)
+//                .build();
+//    }
     //this makes json serialization not to include lazy loaded objects
 //    @Bean
 //    public Hibernate5Module hibernate5Module() {
