@@ -2,12 +2,13 @@ package com.ogoma.hr_core.boundaries.access_control.repositories;
 
 import com.ogoma.hr_core.boundaries.access_control.entities.PasswordReset;
 
+import com.ogoma.hr_core.boundaries.performance.promotion.repository.BaseRepo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PasswordResetRepository extends JpaRepository<PasswordReset,Long> {
+public interface PasswordResetRepository extends BaseRepo<PasswordReset> {
 
     @Nullable
     PasswordReset findByToken(String token);
