@@ -1,9 +1,10 @@
 package com.ogoma.hr_core.boundaries.access_control.repositories;
 
 import com.ogoma.hr_core.boundaries.access_control.entities.Permission;
+import com.ogoma.hr_core.boundaries.performance.promotion.repository.BaseRepo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PermissionsRepository extends JpaRepository<Permission,Long> {
+public interface PermissionsRepository extends BaseRepo<Permission> {
 }
