@@ -3,5 +3,5 @@ package com.ogoma.hr_core.boundaries.documents.controllers;
 import org.springframework.stereotype.Controller;
 
 @Controller
-public class DocumentsController {
+public class CloudObjectsController {
 }
