@@ -1,9 +1,9 @@
 package com.ogoma.hr_core.boundaries.documents.repositories;
 
-import com.ogoma.hr_core.boundaries.documents.entities.Document;
-import org.springframework.data.jpa.repository.JpaRepository;
+import com.ogoma.hr_core.boundaries.documents.entities.CloudObject;
+import com.ogoma.hr_core.boundaries.performance.promotion.repository.BaseRepo;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentsRepository extends JpaRepository<Document,Long> {
+public interface DocumentsRepository extends BaseRepo<CloudObject> {
 }
