@@ -1,6 +1,7 @@
 package com.ogoma.hr_core.boundaries.hr.employee_management.repositories;
 
 import com.ogoma.hr_core.boundaries.hr.employee_management.entities.Designation;
+import com.ogoma.hr_core.boundaries.performance.promotion.repository.BaseRepo;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -11,7 +12,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DesignationRepository extends JpaRepository<Designation, Long> , JpaSpecificationExecutor<Designation> {
+public interface DesignationRepository extends BaseRepo<Designation>{
     Page<Designation> findAllByDeletedFalse(Pageable pageable);
     @Query(value = "select d from Designation d join fetch Department dep")
     List<Designation> findAllByDepartmentId(Long departId);
