@@ -1,9 +1,10 @@
 package com.ogoma.hr_core.boundaries.hr.leave_management.repositories;
 
 import com.ogoma.hr_core.boundaries.hr.leave_management.entities.LeaveType;
+import com.ogoma.hr_core.boundaries.performance.promotion.repository.BaseRepo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeaveTypesRepository extends JpaRepository<LeaveType, Long> {
+public interface LeaveTypesRepository extends BaseRepo<LeaveType> {
 
 
 }
