@@ -1,10 +1,11 @@
 package com.ogoma.hr_core.boundaries.performance.training.repository;
 
+import com.ogoma.hr_core.boundaries.performance.promotion.repository.BaseRepo;
 import com.ogoma.hr_core.boundaries.performance.training.entities.Trainer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TrainersRepository extends JpaRepository<Trainer,Long>, JpaSpecificationExecutor<Trainer> {
+public interface TrainersRepository extends BaseRepo<Trainer> {
 }
