@@ -1,6 +1,6 @@
-package com.ogoma.hr_core.integration;
-
-public class SmtpEmailSenderTest {
-
-
-}
+//package com.ogoma.hr_core.integration;
+//
+//public class SmtpEmailSenderTest {
+//
+//
+//}
